Overload Math::FitsInDWord to handle int32 and int64 separately

+++ lib/Backend/i386/EncoderMD.cpp
@@ -1236,7 +1236,7 @@ modrm:
             bool writeImm = true;
             if (src2 &&src2->IsIntConstOpnd())
             {
-                valueImm = src2->AsIntConstOpnd()->GetImmediateValue();
+                valueImm = (uint)src2->AsIntConstOpnd()->GetImmediateValue();
             }
             else
             {
